Prepare to write some tests for the proxy endpoint
* Move the Betamax serialisers into conftest.py
* Get the existing tests passing
* Use requests.Session in the proxy() endpoint to enable betamax testing
