+++ src/templates/_reading.html
@@ -2,7 +2,9 @@
   <div class="info">
     <div class="book-preview">
       <div class="book-thumbnail">
+        {% if reading.book.image_url %}
         <img src="{{ url_for('proxy', url=reading.book.image_url | quote_plus) }}">
+        {% endif %}
       </div>
       <div class="book-metadata">
         <h3 class="book-title">
